Control Bar Style Changes
Removed the left  and right margins from the tool bar, now that it's the same color as the background and doesn't need to look "set in" from it's container. Noticed that not having a "margin-left" rule made things break, so I put it back and set the margin to zero.