Update Template ID/Class
Updated the map's Handlebars template to use the new ID, and modified class name, that gios-map.js was set to use. This was part of avoiding some weird conflict with the main ASU header.

Also copied a new version of gios-map.js into the project, with those same updated ID/Class changes.