Remove Admin Page
The default Honeycomb plugin has a sample admin page. Although I did mess around with it (and would like to update it so that we can store map colors and options in the database), it doesn't really do anything, so I removed it from the register() method here so that it won't actually be part of the plugin.