Move Disclaimer into a Shortcode Attribute
Changed the way the text disclaimer below the map is handled. It's still part of the Handlebars template, but now you can toggle the <div> on and off through a shortcode attribute called 'disclaimer'.

I also updated the shortcode method itself to avoid use of our own shortcode handling functions and use Wordpress's built-in shortcode_atts() method. This provides us with the same features (avoiding empty arrays, providing default values), and lets us hook into filters. In this case, I wanted to filter the value to convert it to its boolean value according to Wordpress.