Country Name CSS Changes
Changed some of the styling, per Justin's request, on the country name shown above the map after you select a country. It is now a larger font size, and bold.

Also changed the size of the map controls a bit, making them larger on smaller screens - presuming that smaller screens are more likely to be touch screens and need finger-sized butons.