Update shortcode_atts() Method
Turns out that the shortcode_atts() method really wants you to provide default attributes as strings. Updated the code so that the default value for 'disclaimer', our only attribute, as the string 'false', instead of a boolean.

We convert that value to a boolean on the next line, so we can pass it to the template file (which wants a boolean, and NOT a string!); we just need it to be a string initially.